2025-01-17T20:57:00
http://github.com/KenAdeniji/WordEngineering/blob/main/IIS/WordEngineering/2018-05-03Correspondence/2025-01-17T2057AdeyemoAdewuyiAdeniji_-_WeAreSpecializePeopleDontThinkWeAre...WeSpecializeInTeachingSpecializeLesson.txt

We are specialize people, don't think we are... we specialize in teaching specialize lesson.
My adopted father said that in the interval prior to him attending university, he was in a cave in the east.
04:55 file.
My adopted father was probably either lying or sitting down, and his captor was to his north-east and was war-hunting-dressed in a scantily smaller than a kilt, tang.
The evil image setting made me jittery; therefore, my words are jumbled up.

2016-06-23...2025-01-17T20:57:00 http://en.wikipedia.org/wiki/Brexit
	3130 days (8 biblical years, 8 biblical months, 10 days) (8 years, 6 months, 3 weeks, 4 days)

2024-03-20...2025-01-17T20:57:00 In-N-Out Burger. 32060 Union Landing, Store #156. Union City, California (CA) 94587.
	303 days (10 biblical months, 3 days) (9 months, 4 weeks)

2024-04-14...2025-01-17T20:57:00 Eric Menjipar of 99 Ranch Market.
	278 days (9 biblical months, 8 days) (9 months, 3 days)

http://www.ancestry.com/name-origin?surname=menjivar
Menjivar Family History
Menjivar Surname Meaning

Basque: Castilianized form of an unidentified surname probably a topographic name with the second element ibar ‘meadow riverbank valley’.

http://crestsandarms.com/pages/menjivar-family-crest-coat-of-arms
Menjivar is a Spanish surname of Basque origin, derived from the word "mendi" meaning mountain and "ibar" meaning valley. It is believed to have originated as a locational surname for someone who lived near a mountain valley.

2025-02-09T22:47:00
You rest your growth.
Alphabet Sequence
293
Genesis 11:26, 1 Kings 2, Micah 3, Revelation 6:15

Moreover thou knowest also what Joab the son of Zeruiah did to me, and what he did to the two captains of the hosts of Israel, unto Abner the son of Ner, and unto Amasa the son of Jether, whom he slew, and shed the blood of war in peace, and put the blood of war upon his girdle that was about his loins, and in his shoes that were on his feet.
1 Kings 2:5