Scene 1: I was walking towards Newark, California (CA) 94560. 12:27 microsoft windows operating system, microsoft sql server management studio, tab grid falls, drops. Near the intersection of Paseo Padre Parkway and Ardenwood Boulevard, I again, on the second occasion, saw a woman who was either wearing a green or an orange colored bright-colored vest.
Scene 2: I arrived at my destination, 6505 Kaiser Drive. Fremont, California (CA) 94555-3614. I saw a man that I identified by the name of Kennedy who was sitting in the center-west, and he was facing east. Kennedy was surprised and confrontational to see me so early, and Kennedy asked when is our appointment? I said 9 o'clock. I tried to sit in the center-east seat 12:52 facing and opposite Kennedy. 12:58 With the military replacement... who is next? 13:00 right ear. When do we arrive at a sequence... as God? Machine Learning, Deep Learning. A royal family?