The Scripture View's BookTitle Column
The SQL statement

dbo.udf_BookTitle(BookID)

It is for determining this computed column. As expected, there is a correlation between the BookID column, and its corresponding BookTitle column; it progresses from Genesis to Revelation. Because SQL does not support arrays, the author chose to write an SQL CLR C# function for determining the BookTitle when passed the BookID. Although C# supports Design by contract assertions, the author only checks BookID range validity, and returns NULL, if the argument does not fall within this range. The author could throw exceptions, but the author does not know the side effects nor the full ramifications. Instead of writing and determining the book title using C#, an alternative is to use a database table. A table with two columns, BookID and BookTitle, will store and make extractable the sixty-six books. For example, the first book of the New Testament, the 40th book, spelling is Matthew or Mathew, double tt versus single t. An improvement on the current implementation is to use soundex to decipher the book title and the corresponding BookID.