ATS-17119 Request to add an organization from Ken Adeniji Camille Goudeseune<jira@access-ci.atlassian.net> ​ You​ —-—-—-— Reply above this line.
Camille Goudeseune commented:

This is a web page, not an organization such as a university or corporation.
Instead, when you register for ACCESS, for your organization just choose "individual/unaffiliated."

Camille Goudeseune changed the status to Resolved.

Camille Goudeseune resolved this as Won't Do.

How was our service for this request?
☆

Very poor
	☆

Poor
	☆

Average
	☆

Good
	☆

Very good

View request · Turn off this request's notifications

This is shared with kenadeniji@hotmail.com.