2025-06-29T11:28:00...2025-06-29T12:17:00
http://github.com/KenAdeniji/WordEngineering/blob/main/IIS/WordEngineering/2018-05-03Correspondence/2025-06-29T1128ShowerSponge_-_WhoPersonally...UseMe_-_DonaldTrump_MakeAmericaGreatAgain_-_FruitOfTheLoom_GreenElSalvador_-_HeIsAskingForAnEffortOnYourPart_-_AlbertsonsLuckyAlexis.txt

2025-06-29T10:59:00 Who personally... use Me?
    In the shower, I held the sponge.

I asked the customers of Albertsons Lucky, Charter Square, who Alexis was attending to as a clerk. Where are you from?
The husband of the customers asked in response, "Where are you from"?
I answered I am Nigerian, but I used to live in Australia.
The husband partner of the customers responded, "We are not from around here".
    The 2 probably Jewish men at Wienerschnitzel answered the same.
The husband partner of the customers opened a soft drink, soda, and drank from it. The husband of the customers gave a drink to the wife of the customers.
Inside the bakery, the wife of Ray, Agnes, was standing beside the donuts, and I asked for some donuts. She said "they are not cooked yet".

Use-case title: The Budding of Aaron's Rod (Numbers 17).

I remembered the slogan of President Donald Trump, Make America Great Again (MAGA).
    When I was wearing the Fruit of the Loom. Platinum. L/G. 100% cotton. Made in El Salvador. Do not use fabric softener. R/N 13705. DJLA08. Green t-shirt.
    2025-06-29T11:05:00    He is asking... for an effort... on your part.

2025-06-29T11:37:00 Use-case extension: Re-use Microsoft SQL Server Management Studio editor for grid editing, if the script 2025-06-29T12:10:00 does not have to be Transact-SQL compliant.