AutoReply - Thank you for alerting us to a suspected hoax Hoax Mailbox<Hoax@cba.com.au> ​ You​
Thank you for alerting us to a suspected hoax email or SMS. Your feedback is greatly appreciated.

The information you have provided will assist us in removing malicious websites as quickly as possible and helps us keep our customers and communities safe.

If you are concerned about a payment you’ve made or that you need to dispute, please message us in the CommBank app or call 13 2221 8am – 8pm (AEST). If calling from overseas, contact your local International Operator and ask them to reverse charges to +612 9999 3283.

Remember, we will never send you an email or SMS asking for banking information like your NetBank Client ID, password, or NetCode; or include a link to login directly from an email or SMS.

We’re here to help

To learn more about protecting yourself from scams and fraud, please visit commbank.com.au/safe or message us in the CommBank app.

Yours sincerely

The CommBank Team

This is an automatically generated email, due to the volume of emails received by this inbox. Please do not reply.

************** IMPORTANT MESSAGE **************
This e-mail message is intended only for the addressee(s) and contains information which may be confidential.
If you are not the intended recipient please advise the sender by return email, do not use or disclose the contents, and delete the message and any attachments from your system. Unless specifically indicated, this email does not constitute formal advice or commitment by the sender or the Commonwealth Bank of Australia (ABN 48 123 123 124 AFSL and Australian credit licence 234945) or its subsidiaries.
We can be contacted through our web site which can be found by searching CommBank in a browser. If you no longer wish to receive commercial electronic messages from us, please reply to this e-mail by typing Opt Out in the subject line.
**************************************************